refactor(network): introduce custom BitcoinNetwork for store
The `bitcoin::Network` enum from the `rust-bitcoin` crate does not implement `schemars::JsonSchema`, which is required for generating CosmWasm contract schemas.

This commit introduces a new `BitcoinNetwork` enum within the `btc-light-client` contract that mirrors the structure of `bitcoin::Network` but derives the necessary traits for schema generation (`JsonSchema`, `Serialize`, `Deserialize`).

All usages of the previous `babylon_bitcoin::Network` (which was a re-export of `bitcoin::Network`) have been replaced with the new `btc_light_client::state::BitcoinNetwork` enum across the affected contracts (`babylon`, `btc-staking`, `btc-light-client`).

This change ensures that our contract messages and state can be correctly serialized and have their schemas generated, while still allowing for easy conversion to the underlying `bitcoin::Network` type for use with the `rust-bitcoin` library.
