+++ contracts/babylon/src/state/config.rs
@@ -6,7 +6,7 @@ pub(crate) const CONFIG: Item<Config> = Item::new("config");
 
 #[cw_serde]
 pub struct Config {
-    pub network: babylon_bitcoin::Network,
+    pub network: btc_light_client::state::BitcoinNetwork,
     pub babylon_tag: Vec<u8>,
     pub btc_confirmation_depth: u32,
     pub checkpoint_finalization_timeout: u32,
