+++ contracts/btc-finality/src/multitest/suite.rs
@@ -12,7 +12,7 @@ use babylon_apis::finality_api::{IndexedBlock, PubRandCommit};
 use babylon_apis::{btc_staking_api, finality_api, to_bech32_addr, to_canonical_addr};
 use babylon_bindings::BabylonMsg;
 use babylon_bindings_test::BabylonApp;
-use babylon_bitcoin::Network;
+use btc_light_client::state::BitcoinNetwork;
 
 use btc_staking::msg::{
     ActivatedHeightResponse, AllPendingRewardsResponse, FinalityProviderInfo, PendingRewards,
@@ -122,7 +122,7 @@ impl SuiteBuilder {
                 contract_code_id,
                 owner.clone(),
                 &babylon_contract::msg::contract::InstantiateMsg {
-                    network: Network::Testnet,
+                    network: BitcoinNetwork::Testnet,
                     babylon_tag: "01020304".to_string(),
                     btc_confirmation_depth: 1,
                     checkpoint_finalization_timeout: 1,
